+++ cg/meta/demultiplex/demux_post_processing.py
@@ -130,7 +130,7 @@ class DemuxPostProcessingAPI:
 
     def finish_all_flow_cells_temp(self) -> bool:
         """Finish all flow cells that need it."""
-        flow_cell_dirs = self.demux_api.get_all_demultiplexed_flow_cell_dirs()
+        flow_cell_dirs = self.get_all_demultiplexed_flow_cell_dirs()
         is_error_raised: bool = False
         for flow_cell_dir in flow_cell_dirs:
             try:
@@ -154,6 +154,15 @@ class DemuxPostProcessingAPI:
             flow_cell_run_dir=self.demux_api.run_dir,
             store=self.status_db,
         )
+
+    def get_all_demultiplexed_flow_cell_dirs(self) -> List[Path]:
+        """Return all demultiplex flow cell out directories."""
+        demultiplex_flow_cells: List[Path] = []
+        for flow_cell_dir in self.demux_api.out_dir.iterdir():
+            if flow_cell_dir.is_dir():
+                LOG.debug(f"Found directory {flow_cell_dir}")
+                demultiplex_flow_cells.append(flow_cell_dir)
+        return demultiplex_flow_cells
 
 
 class DemuxPostProcessingHiseqXAPI(DemuxPostProcessingAPI):
@@ -270,7 +279,7 @@ class DemuxPostProcessingHiseqXAPI(DemuxPostProcessingAPI):
 
     def finish_all_flow_cells(self, bcl_converter: str) -> None:
         """Loop over all flow cells and post process those that need it."""
-        for flow_cell_dir in self.demux_api.get_all_demultiplexed_flow_cell_dirs():
+        for flow_cell_dir in self.get_all_demultiplexed_flow_cell_dirs():
             self.finish_flow_cell(
                 bcl_converter=bcl_converter,
                 flow_cell_name=flow_cell_dir.name,
@@ -449,5 +458,5 @@ class DemuxPostProcessingNovaseqAPI(DemuxPostProcessingAPI):
 
     def finish_all_flow_cells(self, bcl_converter: str) -> None:
         """Loop over all flow cells and post-process those that need it."""
-        for flow_cell_dir in self.demux_api.get_all_demultiplexed_flow_cell_dirs():
+        for flow_cell_dir in self.get_all_demultiplexed_flow_cell_dirs():
             self.finish_flow_cell(flow_cell_name=flow_cell_dir.name, bcl_converter=bcl_converter)
