+++ tests/meta/demultiplex/test_demux_post_processing.py
@@ -567,3 +567,19 @@ def test_post_processing_of_flow_cell(
     )
 
     assert delivery_path.exists()
+
+
+def test_get_all_demultiplexed_flow_cell_out_dirs(
+    demultiplex_context: CGConfig, demultiplexed_flow_cell_finished_working_directory: Path
+):
+    """Test returning all flow cell directories from the demultiplexing run directory."""
+    # GIVEN a demultiplex flow cell finished output directory that exist
+
+    # GIVEN a demultiplex context
+    demux_api: DemuxPostProcessingAPI = DemuxPostProcessingAPI(config=demultiplex_context)
+
+    # WHEN calling get_all_demultiplexed_flow_cell_dirs
+    demultiplex_flow_cell_dirs: List[Path] = demux_api.get_all_demultiplexed_flow_cell_dirs()
+
+    # THEN the demultiplexed flow cells run directories should be returned
+    assert demultiplex_flow_cell_dirs[0] == demultiplexed_flow_cell_finished_working_directory
